Begin reorganising repo and commit work on new syscall interface
Uppercase in directory paths are a bit annoying to traverse and the root
directory of the repo had too many files/directories. A bit late of a
change but better late than never.

New work has begun on a vDSO (see Kernel/hiraku) and a new syscall
interface. This was initially just an experimental thing and commits
were not made, hence this has all been lumped into one awful, huge commit